im模块完善2 完善业务消息发送链路

+++ live-im-interface/src/main/java/org/example/constant/ImMessageCacheGap.java
@@ -2,6 +2,6 @@ package org.example.constant;
 
 public class ImMessageCacheGap {
 
-    public final static double VALUE = 30;
+    public final static int VALUE = 30;
 
 }
